Move legacy survey writer to survey utils
